moved sse functionality to features folder
