+++ src/features/sse/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@features/sse",
+  "private": true,
+  "main": "./index.ts",
+  "types": "./index.ts"
+}
