+++ src/features/sse/service.ts
