+++ src/features/sse/types.ts
