--- src/lib/sse/index.ts
@@ -1,9 +1,0 @@
-export * from "./service";
-export type {
-  SSEConnection,
-  SSEEvent,
-  SSEConnectionOptions,
-  SSEBroadcastOptions,
-  SSEStats,
-  SSEUserStats,
-} from "./types";
